Introduce Degenerator.Code to handle common code manipulation
